refactor: move `MockPhotometer` into base `Photometer` module
Since the mock photometer should be a strict implementation of the ABC, and isn't related to any hardware, I think it makes the most sense to have it in `hrl/photometer/photomer.py` directly...
